ratelimit: signed check for pipe number in rpc_set_queue()
- rpc can reand signed value

(cherry picked from commit ed034d4d9096d8c77269828c8efdafb26eac5372)

+++ modules/ratelimit/ratelimit.c
@@ -1189,7 +1189,7 @@ static void rpc_set_queue(rpc_t *rpc, void *c) {
 
 	if (rpc->scan(c, "dSd", &queue_no, &method, &pipe_no) < 3) return;
 
-	if (pipe_no >= MAX_PIPES || pipe_no < 0) {
+	if (pipe_no >= MAX_PIPES || (int)pipe_no < 0) {
 		LM_ERR("Invalid pipe number: %d\n", pipe_no);
 		rpc->fault(c, 400, "Invalid pipe number");
 		return;
